Update to notes and evidence upload migrations

+++ database/migrations/2021_03_29_075317_create_student_evidence_table.php
@@ -16,9 +16,9 @@ class CreateStudentEvidenceTable extends Migration
     {
         Schema::create('uploads', function (Blueprint $table) {
             $table->id('id');
-            $table->unsignedInteger('student_id');
-            $table->unsignedInteger('staff_id');
-            $table->unsignedInteger('cohort_id');
+            $table->unsignedBigInteger('student_id');
+            $table->unsignedBigInteger('staff_id');
+            $table->unsignedBigInteger('cohort_id');
             $table->json('fileUploads')->nullable();
             $table->json('comments');
             $table->json('urls')->nullable();
